NETOBSERV-602 Fix computed metrics
All metrics are now derived from the same backend query which is
"sum by(...) (rate(...))" in logql. This provides the Bytes-per-second
(or packets-per-second) metric.

- Rate average: datapoints are then averaged in UI to display edge labels
- Rate max: datapoints are then maxed in UI to display edge labels
- Total: this is "avg x rangeInSeconds"
- Removed the previous "rate" metric that was confusing (it was actually
  the ratio of number of flows (not bytes/packets) versus total flows
queried; it doesn't look so interesting to compute; maybe another
interesting one could be the traffic distribution between peers (in
bytes), but it raises some UI challenges (display 2 labels per edge, one
on incoming and one on outgoing?), so I prefer to not try doing that in
the same PR)
- Added "Latest" label: show the latest rate from the considered time
  window.

Step and interval are now computed from the UI: roughly,
interval=timerange/10 and step=interval/2. Step is necessary to
know in UI to normalize some of the resulting metrics (filling up missing
datapoints for avg)

Fix chart display, tests